requirements.txt
torch==2.4.1 
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2 
scipy==1.10.1       
transformers==4.46.3 
peft==0.13.2         
accelerate==0.26.0   
fair-esm==2.0.0      
huggingface-hub==0.29.1 
safetensors==0.5.3  
tokenizers==0.20.3   
biopython==1.83 
einops==0.8.1   
tqdm==4.67.1 
joblib==1.4.2 
typing_extensions==4.13.2 
matplotlib==3.7.5
seaborn==0.13.2
logomaker==0.8.7
networkx==3.1
nltk==3.9.1
numba==0.58.1
pynndescent==0.5.13
umap-learn==0.5.7   
bitsandbytes==0.45.5 
pillow==10.4.0       
torchaudio==0.8.0
torchvision==0.9.0
